Trait search::DepthFirstTree

pub trait DepthFirstTree where
    Self: Sized, {
    type Step: Sized;
    type Output;
    fn next_steps(&mut self) -> Vec<Self::Step>;

    fn should_prune(&mut self) -> Pruning;

    fn apply_step(&mut self, node: &Self::Step);

    fn revert_step(&mut self, node: &Self::Step);

    fn output(&mut self) -> Option<Self::Output>;

    fn into_iter(self) -> DepthFirstSearcher<Self> { ... }

    fn run_search(self) -> Self { ... }
}

A trait for a tree which is to be traversed, depth-first, in the pursuit of nodes which satisfy a particular condition.

Concepts

There are two concepts that are important in understanding how to use this trait - that of tree "state", and that of tree "steps".

The "state" of a tree is an umbrella term encompassing all the information held about a particular position in the tree - including any values associated with the position and any internal variables that are used to aid the search.

A tree "step" is a small structure - ideally one that can be constructed very cheaply - which contains instructions about how to move between parent/child tree states.

Usage

To define a tree, you must provide the following:

A structure holding the tree state, which you will implement DepthFirstTree for

An associated type Step which represents the tree "steps"

An associated type Output which represents the outputs of the search

Implementations of the required methods of this trait:

next_steps must return the steps that it is valid to take from the current state

should_prune must decide whether to prune the tree at the current state

apply_step must define how to use a step to go from parent to child state

revert_step must define how to use a step to go from child back to parent state

output must take the current state of the tree, and return Some(value) if the current state meets the condition.

Examples

In this example, we will search for all numbers having at most 3 digits, with all digits being odd, and the whole number being divisible by 13.

We will utilise a tree search, with child states coming from the parent by appending a single digit to the parent state.

Our tree state will be the current number being examined, the current number of digits, and the maximum number of digits that a solution should have (in this case, 3).

Our steps will simply contain the digit to be appended when going from parent to child.

use search::{DepthFirstTree, Pruning};

struct Step {
    next_digit: u32,
}

struct Tree {
    value: u32,
    number_of_digits: usize,
    max_digits: usize,
}

impl DepthFirstTree for Tree {
    type Step = Step;
    type Output = u32;

    /// All possible choices for the next digit, which must be odd, as per our constraints.
    fn next_steps(&mut self) -> Vec<Step> {
        [1, 3, 5, 7, 9].iter().map(|&digit| { Step { next_digit: digit } }).collect()
    }

    /// Prune the tree if we have reached the maximum number of digits.
    fn should_prune(&mut self) -> Pruning {
        if self.number_of_digits == self.max_digits {
            Pruning::Below
        } else {
            Pruning::None
        }
    }

    /// Append the next digit to the end of the current value.
    fn apply_step(&mut self, step: &Step) {
        self.value = 10 * self.value + step.next_digit;
        self.number_of_digits += 1;
    }

    /// Remove the last digit from the end of the current value.
    fn revert_step(&mut self, step: &Step) {
        self.value /= 10;
        self.number_of_digits -= 1;
    }

    /// Check if the value is divisible by 13,and output it if so.
    fn output(&mut self) -> Option<u32> {
        if self.value > 0 && self.value % 13 == 0 {
            Some(self.value)
        } else {
            None
        }
    }
}

impl Tree {
    fn with_max_digits(max_digits: usize) -> Tree {
        Tree { value: 0, number_of_digits: 0, max_digits: max_digits }
    }
}

let numbers: Vec<_> = Tree::with_max_digits(3).into_iter().collect();
assert_eq!(numbers, vec![117, 13, 195, 351, 377, 39, 533, 559, 715, 793, 91, 975]);

Associated Types

type Step: Sized

type Output

Loading content...
Required methods

fn next_steps(&mut self) -> Vec<Self::Step>

Returns the steps that can be taken from the current state of the tree.

fn should_prune(&mut self) -> Pruning

Returns whether to prune the tree at the current state.

fn apply_step(&mut self, node: &Self::Step)

Update the tree-wide state as a result of applying the given step.

fn revert_step(&mut self, node: &Self::Step)

Update any tree-wide state by reverting the given step.

fn output(&mut self) -> Option<Self::Output>

Maps the internal state of the tree to the actual form required for output.

Returns Some(value) if the current state satisfies the condition, and None otherwise.

Loading content...
Provided methods

Important traits for DepthFirstSearcher<T>
impl<T: DepthFirstTree> Iterator for DepthFirstSearcher<T> type Item = T::Output;

fn into_iter(self) -> DepthFirstSearcher<Self>

An iterator over the nodes of the tree which meet the condition.

fn run_search(self) -> Self

Run this tree's complete search syncrhonously

Loading content...