search
Module types

This module defines the types that we export across the UNIFFI interface.

Structs

RefinedSearchConfig

Details of the search engines to display to the user, generated as a result of processing the search configuration.

SearchEngineDefinition

A definition for an individual search engine to be presented to the user.

SearchEngineUrl

Defines an individual search engine URL.

SearchEngineUrls

The URLs associated with the search engine.

SearchUrlParam

Parameter definitions for search engine URLs. The name property is always specified, along with one of value, experiment_config or search_access_point.

SearchUserEnvironment

The user’s environment that is used for filtering the search configuration.

Enums

SearchApplicationName

The list of possible application names that are currently supported.

SearchDeviceType

SearchEngineClassification

The list of acceptable classifications for a search engine.

SearchUpdateChannel

The list of possible update channels for a user’s build. Use default for a self-build or an unknown channel.